pyproject.toml
[project]
name = "status"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
 "arxiv-base @ git+https://github.com/arXiv/arxiv-base.git@master"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"